Render slides based on child components
Allow the carousel to cycle through its child components (1 child = 1 slide) instead of the `renderItem()` and `items` prop. 

Initially, slides were rendered based on both an `item` and a `renderItem() prop. The problem with that pattern is that it only allow for "dumb" components to be used as a slide: any more sophisticated component containing references to the parent states/props (like ListViews with a dataSource coming from the parent component) would not work (because carousel is not aware of the parent state/prop).
One approach would have been to just pass all props and states in the carousel component with `{...props}` but I found that the better approach is to invite users to list each slide as a child of the carousel component and then rely on cycling though each child to know what to display. This way we avoid passing more props than necessary, and we don't rely on a mapping done in the component it self: we are in charge of displaying each slides (1 child of the carousel = 1 slide).

!!! As a downside: this has the bad effect to break previous versions... 
In fact, now the props `items` and `renderItem` are deprecated.
And one has to explicitly feed each slides it has to display, like this:

```
<Carousel>
   <View key={0} />   // Slide 1
   <View key={1} />   // Slide 2
   <View key={2} />   // Slide 3
</Carousel>
```